requirements.txt
discord.py>=1.3.3,<2.0.0
gspread-asyncio>=1.1.0,<2.0.0
oauth2client
srp>=1.0.16<2.0.0
aiohttp_socks
Pillow>=7.1.2,<8.0.0
beautifulsoup4>=4.9.1,<5.0.0
lxml>=4.5.1,<5.0.0
pymongo[srv]>=3.10.1,<4.0.0
motor>=2.1.0,<3.0.0